#14: Restructuring util package
