+++ .gitignore
@@ -1,2 +1,3 @@
 atlas.yaml
 dist/
+go.sum
