+++ cmd/repo.go
@@ -3,7 +3,7 @@ package cmd
 import (
 	"os/exec"
 
-	"github.com/ahstn/atlas/util"
+	"github.com/ahstn/atlas/pkg/util"
 	"github.com/urfave/cli"
 	emoji "gopkg.in/kyokomi/emoji.v1"
 )
