+++ go.mod
@@ -4,9 +4,16 @@ require (
 	github.com/apex/log v1.0.0
 	github.com/briandowns/spinner v0.0.0-20180529140538-1567cd82701b
 	github.com/fatih/color v1.7.0
+	github.com/mattn/go-colorable v0.0.9 // indirect
+	github.com/mattn/go-isatty v0.0.3 // indirect
 	github.com/pkg/errors v0.8.0
 	github.com/urfave/cli v1.20.0
+	gopkg.in/VividCortex/ewma.v1 v1.1.1 // indirect
 	gopkg.in/cheggaaa/pb.v2 v2.0.6
+	gopkg.in/fatih/color.v1 v1.7.0 // indirect
 	gopkg.in/kyokomi/emoji.v1 v1.5.1
+	gopkg.in/mattn/go-colorable.v0 v0.0.9 // indirect
+	gopkg.in/mattn/go-isatty.v0 v0.0.3 // indirect
+	gopkg.in/mattn/go-runewidth.v0 v0.0.2 // indirect
 	gopkg.in/yaml.v2 v2.2.1
 )
